refactor: rename particles to particle
