+++ expertsystem/__init__.py
@@ -7,13 +7,13 @@ etc.).
 
 The `expertsystem` consists of three main components:
 
-#. `particles`: a stand-alone submodule with which one can investigate specific
+#. `particle`: a stand-alone submodule with which one can investigate specific
    quantum properties of `.Particle` instances (see :doc:`/usage/particles`).
 
 #. `reaction`: the core of the `expertsystem` that computes which transitions
    (represented by a `.StateTransitionGraph`) are allowed between a certain
    initial and final state. Internally, the system propagates the quantum
-   numbers defined by `particles` through the `.StateTransitionGraph`, while
+   numbers defined by `particle` through the `.StateTransitionGraph`, while
    satisfying the rules define by the :mod:`.conservation_rules` module.
 
 #. `amplitude`: a collection of tools to convert the `.StateTransitionGraph`
@@ -23,17 +23,17 @@ The `expertsystem` consists of three main components:
 
 Finally, the `.ui` module glues these modules together through facade functions
 and the `.StateTransitionManager`, while the `.io` module provides tools that
-can read and write the objects of `particles`, `reaction`, and `amplitude`.
+can read and write the objects of `particle`, `reaction`, and `amplitude`.
 """
 
 
 __all__ = [
     "amplitude",
     "io",
-    "particles",
+    "particle",
     "reaction",
     "ui",
 ]
 
 
-from . import amplitude, io, particles, reaction, ui
+from . import amplitude, io, particle, reaction, ui
