+++ expertsystem/reaction/quantum_numbers.py
@@ -4,14 +4,14 @@ The `expertsystem` is strictly typed (enforced through :doc:`mypy
 <mypy:index>`. This module bundles structures and definitions that don't serve
 as data containers but only as type hints. `.EdgeQuantumNumbers` and
 `.NodeQuantumNumbers` are the main structures and serve as a bridge between the
-:mod:`.particles` and the :mod:`.reaction` module.
+:mod:`.particle` and the :mod:`.reaction` module.
 """
 
 from typing import NewType, Optional, Tuple, Union
 
 import attr
 
-from expertsystem.particles import Parity, Particle
+from expertsystem.particle import Parity, Particle
 
 ParticleWithSpin = Tuple[Particle, float]
