+++ pyproject.toml
@@ -34,3 +34,4 @@ exclude = '''
 profile = "black"
 src_paths = ["expertsystem", "tests"]
 line_length = 79
+known_third_party = "THIRDPARTY,particle"
