+++ tests/unit/io/test_particle_list.py
@@ -2,7 +2,7 @@
 import pytest
 
 from expertsystem import io
-from expertsystem.particles import ParticleCollection
+from expertsystem.particle import ParticleCollection
 
 XML_FILE = "particle_list.xml"
 YAML_FILE = "particle_list.yml"
